Add        Project/BinarySerializerNative/Compile/BufferStream.cpp Modified   Project/NetworkingNative/Compile/Constants.cpp Modified   Project/BinarySerializerManaged/BufferStream.cs Modified   Project/NetworkingManaged/BaseSocket.cs Modified   Project/NetworkingManaged/Constants.cs Modified   Project/Binaries/ASCIISerializerManaged.dll Modified   Project/Binaries/NetworkingManaged.NetCore.dll Modified   Project/Binaries/NetworkingManaged.NetFramework.dll Modified   Project/BinarySerializerNative/BinarySerializerNative.vcxproj.filters Modified   Project/NetworkingNative/NetworkingNative.vcxproj.filters Add        Project/BinarySerializerNative/Include/BufferStream.h Add        Project/BinarySerializerNative/Include/Utilities.h Modified   Project/NetworkingNative/Include/BaseSocket.h Modified   Project/NetworkingNative/Include/Constants.h Modified   Project/NetworkingNative/Include/Enums.h Modified   Project/NetworkingNative/Include/SocketUtilities.h Modified   Project/BinarySerializerNative/BinarySerializerNative.vcxproj Modified   Project/NetworkingNative/NetworkingNative.vcxproj

+++ Project/NetworkingManaged/BaseSocket.cs
@@ -124,7 +124,7 @@ namespace GameFramework.Networking
 			Socket.SendBufferSize = (int)Constants.SEND_BUFFER_SIZE;
 			Socket.ReceiveTimeout = (int)Constants.RECEIVE_TIMEOUT;
 			Socket.SendTimeout = (int)Constants.SEND_TIMEOUT;
-			Socket.Ttl = Constants.TIME_TO_LIVE;
+			Socket.Ttl = (short)Constants.TIME_TO_LIVE;
 
 			SocketUtilities.SetIPv6OnlyEnabled(Socket, false);
 			SocketUtilities.SetChecksumEnabled(Socket, false);
